ILMS-125 - refactor files
