--- src/components/Settings/assets/index.ts
@@ -1,7 +1,0 @@
-import { VideoIcon } from "./VideoIcon";
-import { AudioIcon } from "./AudioIcon";
-import { ScreenIcon } from "./ScreenIcon";
-import { NoVideoIcon } from "./NoVideoIcon";
-import { NoAudioIcon } from "./NoAudioIcon";
-import { NoScreenIcon } from "./NoScreenIcon";
-export { VideoIcon, AudioIcon, ScreenIcon, NoAudioIcon, NoScreenIcon, NoVideoIcon };
