+++ src/components/Settings/icons/index.ts
@@ -1,0 +1,8 @@
+import { VideoIcon } from "./video";
+import { AudioIcon } from "./audio";
+import { NoAudioIcon } from "./noAudio";
+import { ScreenIcon } from "./screen";
+import { NoVideoIcon } from "./noVideo";
+import { NoScreenIcon } from "./noScreen";
+import { CheckIcon } from "./check";
+export { VideoIcon, AudioIcon, ScreenIcon, NoVideoIcon, NoScreenIcon, NoAudioIcon, CheckIcon };
